+++ CHANGES.rst
@@ -17,6 +17,7 @@ Changes
  - launchpad #1142994: url rewrites for vcs polling
  - launchpad #1154673: treeStableTimer scheduler parameter now
    configurable on a per-buildout basis
+ - launchpad #1281136: subfactory for sphinx doc compilation & upload
 
 0.8.1
 -----